Updated queue.h file
